fix: move from nodejs22 into nodejs and remove nodejs22 dir
