+++ nodejs/package.json
@@ -1,6 +1,7 @@
 {
-  "name": "fission-nodejs-runtime",
-  "version": "0.1.0",
+  "name": "fission-nodejs-22-runtime",
+  "version": "1.0.0",
+  "type": "module",
   "author": "Soam Vasani",
   "contributors": [
     {
@@ -12,23 +13,18 @@
       "email": "contact@garyyeap.com"
     }
   ],
-  "description": "NodeJS run container for the fission framework",
+  "description": "Pure ESM Node.js 22 runtime for Fission with multi-entry support, modern dependencies, and enhanced function loading",
   "engines": {
-    "node": ">= 18.0.0"
+    "node": ">= 22.0.0"
   },
   "dependencies": {
-    "co": "~4.6.0",
-    "express": "^4.18.2",
+    "express": "^5.1.0",
     "minimist": "^1.2.8",
-    "morgan": "^1.10.0",
-    "mz": "~2.7.0",
-    "request": "^2.88.2",
-    "request-promise-native": "^1.0.9",
-    "underscore": "^1.13.6",
-    "ws": "^8.14.2"
+    "pino": "^9.4.0",
+    "pino-pretty": "^11.2.2",
+    "ws": "^8.18.3"
   },
-  "overrides": {
-    "tough-cookie": "^4.1.3",
-    "semver": "^7.5.2"
+  "scripts": {
+    "start": "node server.js"
   }
 }
