+++ nodejs/test/test-case-3/package.json
@@ -1,0 +1,3 @@
+{
+  "type": "commonjs"
+}
