+++ nodejs/test/test-case-4/momentExample.js
@@ -1,8 +1,18 @@
-const momentpackage = require("moment");
+// ESM momentExample test case for Node.js 22
+import moment from "moment";
 
-module.exports = async (context) => {
+export default async (context) => {
   return {
     status: 200,
-    body: "Hello " + momentpackage().format(),
+    body: JSON.stringify({
+      message: "Hello from ESM Node.js 22! 🕐",
+      timestamp: moment().format(),
+      moduleType: "ESM",
+      nodeVersion: process.version
+    }),
+    headers: {
+      "Content-Type": "application/json",
+      "X-Module-Type": "ESM"
+    }
   };
-};
+};
