+++ nodejs/test/test-case-4/package.json
@@ -1,9 +1,19 @@
 {
   "name": "fission-nodejs-runtime",
+  "type": "module",
   "engines": {
     "node": ">=7.6.0"
   },
   "dependencies": {
     "moment": "*"
-  }
+  },
+  "version": "1.0.0",
+  "main": "momentExample.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "description": ""
 }
