+++ nodejs/test/test-case-5/package.json
@@ -1,0 +1,3 @@
+{
+  "type": "module"
+}
