+++ nodejs/test/test-case-7/package.json
@@ -1,0 +1,3 @@
+{
+  "type": "module"
+}
