+++ nodejs/test/test-case-7/wordCount.js
