--- nodejs22/Dockerfile
@@ -1,17 +1,0 @@
-ARG NODE_BASE_IMG
-
-FROM node:${NODE_BASE_IMG}
-
-ARG NODE_ENV
-ENV NODE_ENV $NODE_ENV
-
-RUN mkdir -p /usr/src/app
-WORKDIR /usr/src/app
-
-COPY package.json /usr/src/app/
-RUN npm install && npm cache clean --force
-COPY server.js /usr/src/app/server.js
-
-CMD [ "npm", "start" ]
-
-EXPOSE 8888
