--- nodejs22/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "fission-nodejs-22-runtime",
-  "version": "1.0.0",
-  "type": "module",
-  "author": "Soam Vasani",
-  "contributors": [
-    {
-      "name": "Soam Vasani",
-      "email": "soamvasani+1@gmail.com"
-    },
-    {
-      "name": "Gary Yeap",
-      "email": "contact@garyyeap.com"
-    }
-  ],
-  "description": "Pure ESM Node.js 22 runtime for Fission with multi-entry support, modern dependencies, and enhanced function loading",
-  "engines": {
-    "node": ">= 22.0.0"
-  },
-  "dependencies": {
-    "express": "^5.1.0",
-    "minimist": "^1.2.8",
-    "pino": "^9.4.0",
-    "pino-pretty": "^11.2.2",
-    "ws": "^8.18.3"
-  },
-  "scripts": {
-    "start": "node server.js"
-  }
-}
